+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/actions/internal/InternalRuleSetUtil.java
@@ -5,7 +5,6 @@
 
 package net.sourceforge.pmd.eclipse.ui.actions.internal;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
@@ -13,7 +12,6 @@ import java.util.List;
 import java.util.Set;
 import java.util.regex.Pattern;
 
-import net.sourceforge.pmd.lang.document.FileId;
 import net.sourceforge.pmd.lang.rule.Rule;
 import net.sourceforge.pmd.lang.rule.RulePriority;
 import net.sourceforge.pmd.lang.rule.RuleSet;
@@ -88,16 +86,6 @@ public final class InternalRuleSetUtil {
             result.add(Pattern.compile(p));
         }
         return result;
-    }
-
-    public static boolean ruleSetsApplies(List<RuleSet> rulesets, File file) {
-        FileId fileId = FileId.fromPathLikeString(file.toString());
-        for (RuleSet ruleSet : rulesets) {
-            if (ruleSet.applies(fileId)) {
-                return true;
-            }
-        }
-        return false;
     }
 
     public static int countRules(List<RuleSet> rulesets) {